// src/types/Color.ts
export enum Color {
  black = '#0e0e10',
  brown = '#784a2a',
  red = '#9d2825',
  orange = '#e97635',
  yellow = '#eeb32c',
  green = '#6a9742',
  blue = '#317eae',
  violet = '#756a98',
  grey = '#7c888e',
  white = '#e1d8c7',
  gold = '#f9d433',
  silver = '#c0c0c0'
}

export type Indexes = '0' | '1' | '2' | '3' | '4' | '5'